extract pack from bloom service
