+++ service/pack/error.go
@@ -1,0 +1,7 @@
+package pack
+
+import "errors"
+
+var (
+	ErrInvalidType = errors.New("invalid Type")
+)
